numaflow-dra/step1/vertex/pyproject.toml
[tool.poetry]
name = "step1"
version = "1.0.0"
description = ""
authors = ["sesame0224"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pynumaflow = "0.9.1"

python-dotenv = "1.0.1"
opencv-python = "4.10.0.84"
numpy = "2.2.2"
PyYAML = "6.0.2"
pathlib = "1.0.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"